Move common routines to zabob.common for use in all environments.
