+++ houdini/zcommon/pyproject.toml
@@ -8,8 +8,10 @@ authors = [
 ]
 requires-python = ">=3.11"
 dependencies = [
+    "click>=8.2.1",
     "hou>=0.1.0",
-
+    "psutil>=7.0.0",
+    "semver>=3.0.4",
 ]
 
 [build-system]
