+++ houdini/zcommon/src/zabob/common/__init__.py
@@ -9,9 +9,30 @@ from zabob.common.common_paths import (
     ZABOB_ROOT,
     ZABOB_HOME_DIR,
     ZABOB_OUT_DIR,
-    ZABOB_HOUDINI_DATA
+    ZABOB_HOUDINI_DATA,
+    ZABOB_PYCACHE_DIR,
 )
 from zabob.common.timer import timer
+from zabob.common.subproc import (
+    run,
+    capture,
+    spawn,
+    check_pid,
+    CompletedProcess,
+)
+from zabob.common.click_types import (
+    OptionalType, SemVerParamType, OrType, NoneType,
+)
+from zabob.common.common_utils import (
+    _version, Level, LEVELS,
+    DEBUG, INFO, QUIET, SILENT, VERBOSE,
+    environment,
+)
+from zabob.common.find_houdini import (
+    find_houdini_installations,
+    get_houdini,
+    HoudiniInstall,
+)
 
 __all__ = (
     "ZABOB_COMMON_DIR",
@@ -21,5 +42,27 @@ __all__ = (
     "ZABOB_HOME_DIR",
     "ZABOB_OUT_DIR",
     "ZABOB_HOUDINI_DATA",
-    "timer"
+    "ZABOB_PYCACHE_DIR",
+    "timer",
+    "run",
+    "capture",
+    "spawn",
+    "check_pid",
+    "CompletedProcess",
+    "OptionalType",
+    "SemVerParamType",
+    "OrType",
+    "NoneType",
+    "_version",
+    "Level",
+    "LEVELS",
+    "DEBUG",
+    "INFO",
+    "QUIET",
+    "SILENT",
+    "VERBOSE",
+    "environment",
+    "find_houdini_installations",
+    "get_houdini",
+    "HoudiniInstall",
 )
