+++ zabob-modules/src/zabob/core/houdini.py
@@ -15,12 +15,12 @@ import click
 from semver import Version
 
 
-from zabob._find.types import HoudiniInstall
+from zabob.common._find.types import HoudiniInstall
+from zabob.common.find_houdini import get_houdini, list_houdini_installations, show_houdini
+from zabob.common.subproc import run
 from zabob.core.houdini_versions import cli as houdini_cli, download_houdini_installer
-from zabob.core.find_houdini import get_houdini, list_houdini_installations, show_houdini
 from zabob.core.main import main
 from zabob.core.paths import HOUDINI_PROJECTS, ZABOB_HOUDINI_DIR
-from zabob.core.subproc import run
 from zabob.core.hython import hython
 
 @main.group('houdini')
