+++ zabob-modules/src/zabob/core/setup.py
@@ -23,12 +23,13 @@ from tomlkit.toml_file import TOMLFile
 
 if __name__ == '__main__':
     sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
+from zabob.common.common_utils import DEBUG, QUIET, INFO, VERBOSE
+from zabob.common.subproc import capture, run
 from zabob.core.houdini import setup_houdini_venv_from_current
 from zabob.core.node import node_version, node_path  # type: ignore # noqa: E402
 from zabob.core.paths import ALLPROJECTS, ZABOB_CHECKSUMS, ZABOB_ROOT, SUBPROJECTS
-from zabob.core.subproc import capture, run
-from zabob.core.utils import DEBUG, QUIET, INFO, VERBOSE, rmdir
 from zabob.core.main import main
+from zabob.core.utils import rmdir
 
 
 @main.command(name='setup')
